+++ app/slack/templates/messages.py
@@ -273,6 +273,7 @@ class LoginMessage(SlackMessage):
         self._variation = variation
 
         # Have variations of the login message depending on the arguments.
+        ibm = is_ibm_enterprise(enterprise_id)
         block_text = _(
             "In order to use the Straker Translate features, please login. Click this button below;"
         )
@@ -287,7 +288,14 @@ class LoginMessage(SlackMessage):
                 "Connect your account to evaluate the quality of your translation."
             )
         elif variation == self.HUMAN_TRANSLATION:
-            block_text = _("Connect your account to perform human translation.")
+            block_text = (
+                _(
+                    "Human Translation does not require a LanguageCloud login in this "
+                    "workspace."
+                )
+                if ibm
+                else _("Connect your account to perform human translation.")
+            )
         elif isinstance(ray_client, RayClient):
             user_details = f"<{domains.verify}|{ray_client.username}>"
             block_text = _(
@@ -295,50 +303,43 @@ class LoginMessage(SlackMessage):
             )
             if ray_client.sso:
                 block_text = _("Your connected account is: {user_details}")
+        if (
+            ibm
+            and not isinstance(ray_client, RayClient)
+            and variation not in (self.HUMAN_TRANSLATION,)
+        ):
+            block_text = _(
+                "This feature requires a LanguageCloud account provisioned by your "
+                "administrator. Human Translation does not require signing in."
+            )
         msg: list[dict[str, Any]] = [
             {
                 "type": "section",
                 "text": {"type": "mrkdwn", "text": block_text},
             },
         ]
-        if not isinstance(ray_client, RayClient):
-            if is_ibm_enterprise(enterprise_id):
-                msg.append(
-                    {
-                        "type": "actions",
-                        "elements": [
-                            {
-                                "type": "button",
-                                "text": {
-                                    "type": "plain_text",
-                                    "text": _("Direct Login"),
-                                },
-                                "style": "primary",
-                                "action_id": "login_sso",
+        # RAY-81247: IBM Direct Login (SSO mint/link button) removed. Non-IBM
+        # workspaces still use Connect account.
+        if not isinstance(ray_client, RayClient) and not ibm:
+            msg.append(
+                {
+                    "type": "actions",
+                    "elements": [
+                        {
+                            "type": "button",
+                            "text": {
+                                "type": "plain_text",
+                                "text": _("Connect account"),
                             },
-                        ],
-                    }
-                )
-            else:
-                msg.append(
-                    {
-                        "type": "actions",
-                        "elements": [
-                            {
-                                "type": "button",
-                                "text": {
-                                    "type": "plain_text",
-                                    "text": _("Connect account"),
-                                },
-                                "style": "primary",
-                                "url": get_language_cloud_connect_url(
-                                    user_id, team_id, enterprise_id, channel_id
-                                ),
-                                "action_id": "login",
-                            }
-                        ],
-                    }
-                )
+                            "style": "primary",
+                            "url": get_language_cloud_connect_url(
+                                user_id, team_id, enterprise_id, channel_id
+                            ),
+                            "action_id": "login",
+                        }
+                    ],
+                }
+            )
         super().__init__(
             "Connect your account",
             msg,
@@ -2209,50 +2210,53 @@ def get_account_blocks(
             }
         )
     else:
-        account_blocks.append(
-            {
-                "type": "section",
-                "text": {
-                    "type": "mrkdwn",
-                    "text": _(
-                        "In order to use the Straker Translate features, please login. Click this button below;"
-                    ),
-                },
-            }
-        )
-        account_blocks.append(
-            {
-                "type": "actions",
-                "elements": [
-                    (
-                        (
-                            {
-                                "type": "button",
-                                "text": {
-                                    "type": "plain_text",
-                                    "text": _("Connect account"),
-                                },
-                                "style": "primary",
-                                "url": get_language_cloud_connect_url(
-                                    user_id, team_id, enterprise_id, channel_id
-                                ),
-                                "action_id": "login",
-                            }
-                            if not is_ibm
-                            else {
-                                "type": "button",
-                                "text": {
-                                    "type": "plain_text",
-                                    "text": _("Direct Login"),
-                                },
-                                "style": "primary",
-                                "action_id": "login_sso",
-                            }
-                        ),
-                    )
-                ],
-            }
-        )
+        if is_ibm:
+            text = _(
+                "No LanguageCloud account is linked. Human Translation does not "
+                "require signing in; other account features need an administrator "
+                "to provision access."
+            )
+            account_blocks.append(
+                {
+                    "type": "section",
+                    "text": {
+                        "type": "mrkdwn",
+                        "text": text,
+                    },
+                }
+            )
+        else:
+            text = _(
+                "In order to use the Straker Translate features, please login. Click this button below;"
+            )
+            account_blocks.append(
+                {
+                    "type": "section",
+                    "text": {
+                        "type": "mrkdwn",
+                        "text": text,
+                    },
+                }
+            )
+            account_blocks.append(
+                {
+                    "type": "actions",
+                    "elements": [
+                        {
+                            "type": "button",
+                            "text": {
+                                "type": "plain_text",
+                                "text": _("Connect account"),
+                            },
+                            "style": "primary",
+                            "url": get_language_cloud_connect_url(
+                                user_id, team_id, enterprise_id, channel_id
+                            ),
+                            "action_id": "login",
+                        },
+                    ],
+                }
+            )
 
     return account_blocks, text